+++ test/grax/id/counter/dets_test.exs
@@ -1,0 +1,3 @@
+defmodule Grax.Id.Counter.DetsTest do
+  use Grax.Id.Counter.Adapter.TestCase, adapter: Grax.Id.Counter.Dets
+end