+++ test/grax/id/counter/text_file_test.exs
@@ -1,0 +1,3 @@
+defmodule Grax.Id.Counter.TextFileTest do
+  use Grax.Id.Counter.Adapter.TestCase, adapter: Grax.Id.Counter.TextFile
+end